chore: refactor numbers and dots pagination
